refactor: data 레이어 패키지 정리
