+++ app/src/main/java/woowacourse/shopping/data/datasource/local/recent/RecentDao.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.sql.recent
+package woowacourse.shopping.data.datasource.local.recent
 
 import android.content.ContentValues
 import android.content.Context
