+++ app/src/main/java/woowacourse/shopping/data/datasource/remote/product/ProductRetrofitService.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.service
+package woowacourse.shopping.data.datasource.remote.product
 
 import com.example.domain.model.Product
 import retrofit2.Call
@@ -7,6 +7,7 @@ import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 import woowacourse.shopping.data.model.ProductEntity
 import woowacourse.shopping.data.model.toDomain
+import woowacourse.shopping.data.datasource.remote.ServerInfo
 
 class ProductRetrofitService : ProductRemoteDataSource {
