+++ app/src/main/java/woowacourse/shopping/feature/main/MainActivity.kt
@@ -9,16 +9,16 @@ import androidx.databinding.DataBindingUtil
 import androidx.recyclerview.widget.ConcatAdapter
 import androidx.recyclerview.widget.GridLayoutManager
 import woowacourse.shopping.R
-import woowacourse.shopping.data.CartCache
-import woowacourse.shopping.data.CartRemoteRepositoryImpl
-import woowacourse.shopping.data.ProductCacheImpl
-import woowacourse.shopping.data.ProductRemoteRepositoryImpl
-import woowacourse.shopping.data.RecentProductRepositoryImpl
-import woowacourse.shopping.data.TokenSharedPreference
-import woowacourse.shopping.data.service.CartRemoteService
-import woowacourse.shopping.data.service.ProductRetrofitService
-import woowacourse.shopping.data.service.ServerInfo
-import woowacourse.shopping.data.sql.recent.RecentDao
+import woowacourse.shopping.data.datasource.local.cart.CartCache
+import woowacourse.shopping.data.repository.cart.CartRemoteRepositoryImpl
+import woowacourse.shopping.data.datasource.local.product.ProductCacheImpl
+import woowacourse.shopping.data.repository.product.ProductRemoteRepositoryImpl
+import woowacourse.shopping.data.repository.recent.RecentProductRepositoryImpl
+import woowacourse.shopping.data.datasource.local.auth.TokenSharedPreference
+import woowacourse.shopping.data.datasource.remote.cart.CartRemoteService
+import woowacourse.shopping.data.datasource.remote.product.ProductRetrofitService
+import woowacourse.shopping.data.datasource.remote.ServerInfo
+import woowacourse.shopping.data.datasource.local.recent.RecentDao
 import woowacourse.shopping.databinding.ActivityMainBinding
 import woowacourse.shopping.feature.cart.CartActivity
 import woowacourse.shopping.feature.detail.DetailActivity
