+++ app/src/main/java/woowacourse/shopping/feature/serverSetting/ServerSettingActivity.kt
@@ -3,8 +3,8 @@ package woowacourse.shopping.feature.serverSetting
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import woowacourse.shopping.data.TokenSharedPreference
-import woowacourse.shopping.data.service.ServerInfo
+import woowacourse.shopping.data.datasource.local.auth.TokenSharedPreference
+import woowacourse.shopping.data.datasource.remote.ServerInfo
 import woowacourse.shopping.databinding.ActivityServerSettingBinding
 import woowacourse.shopping.feature.main.MainActivity
 import java.util.Base64
