+++ app/src/test/java/woowacourse/shopping/data/ProductRepositoryTest.kt
@@ -3,7 +3,9 @@ package woowacourse.shopping.data
 import com.example.domain.datasource.productsDatasource
 import com.example.domain.model.Product
 import org.junit.Test
-import woowacourse.shopping.data.service.ProductMockWebService
+import woowacourse.shopping.data.datasource.local.product.ProductCacheImpl
+import woowacourse.shopping.data.datasource.remote.product.ProductMockWebService
+import woowacourse.shopping.data.repository.product.ProductRemoteMockRepositoryImpl
 import java.util.concurrent.CountDownLatch
 
 internal class ProductRepositoryTest {
